backend/auth_service/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
gunicorn
psycopg2-binary
prometheus-flask-exporter==0.23.0
